feat(BadgePrimitive): migrate to Tailwind
BREAKING CHANGE: background and foregroundColor props were removed
