+++ docs/src/__examples__/BadgePrimitive/DEFAULT.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import BadgePrimitive from "@kiwicom/orbit-components/lib/primitives/BadgePrimitive";
 import { Stack } from "@kiwicom/orbit-components";
-import defaultTheme from "@kiwicom/orbit-components/lib/defaultTheme";
 import { AirplaneTakeoff } from "@kiwicom/orbit-components/icons";
 
 export default {
@@ -9,10 +8,7 @@ export default {
     <Stack direction="column">
       <BadgePrimitive
         icon={<AirplaneTakeoff />}
-        // TODO: uncomment after Badge is migrated to Tailwind
-        // className="text-white-normal bg-gradient-to-r from-[#fd1d1d] to-[#ffae28]"
-        background="linear-gradient(264.15deg, #FC8D3E 0%, #F16E5B 100%)"
-        foregroundColor={defaultTheme.orbit.colorTextWhite}
+        className="text-white-normal bg-gradient-to-r from-[#fd1d1d] to-[#ffae28]"
       >
         Orbit
       </BadgePrimitive>
