+++ packages/orbit-components/src/primitives/BadgePrimitive/README.md
@@ -16,16 +16,15 @@ After adding import into your project you can use it simply like:
 
 Table below contains all types of the props available in BadgePrimitive component.
 
-| Name            | Type                    | Default | Description                                                          |
-| :-------------- | :---------------------- | :------ | :------------------------------------------------------------------- |
-| background      | `string`                |         | Background of a BadgePrimitive. Can use gradients and images         |
-| foregroundColor | `string`                |         | Foreground color, controling color of a text and icon.               |
-| children        | `React.Node`            |         | The content of the BadgePrimitive.                                   |
-| dataTest        | `string`                |         | Optional prop for testing purposes.                                  |
-| id              | `string`                |         | Set `id` for `BadgePrimitive`                                        |
-| icon            | `React.Node`            |         | The displayed icon on the left.                                      |
-| ariaLabel       | `string`                |         | Adds prop adds `aria-label` to an element, useful for screenreaders. |
-| **carriers**    | [`Carrier[]`](#carrier) |         | The content of the CarrierLogo, passed as array of objects.          |
+| Name         | Type                    | Default | Description                                                                                                         |
+| :----------- | :---------------------- | :------ | :------------------------------------------------------------------------------------------------------------------ |
+| className    |                         |         | Optional className of the component to allow adding classes that apply styles to text, border and background colors |
+| children     | `React.Node`            |         | The content of the BadgePrimitive.                                                                                  |
+| dataTest     | `string`                |         | Optional prop for testing purposes.                                                                                 |
+| id           | `string`                |         | Set `id` for `BadgePrimitive`                                                                                       |
+| icon         | `React.Node`            |         | The displayed icon on the left.                                                                                     |
+| ariaLabel    | `string`                |         | Adds prop adds `aria-label` to an element, useful for screenreaders.                                                |
+| **carriers** | [`Carrier[]`](#carrier) |         | The content of the CarrierLogo, passed as array of objects.                                                         |
 
 ### Carrier
