+++ packages/orbit-components/src/primitives/BadgePrimitive/__tests__/index.test.tsx
@@ -8,9 +8,8 @@ describe("BadgePrimitive", () => {
   it("should have expected DOM output", () => {
     render(
       <BadgePrimitive
+        className="orbit-kek"
         icon={<Sightseeing dataTest="icon" />}
-        background="red"
-        foregroundColor="blue"
         dataTest="test"
         ariaLabel="label"
       >
@@ -18,11 +17,7 @@ describe("BadgePrimitive", () => {
       </BadgePrimitive>,
     );
 
-    expect(screen.getByTestId("test")).toHaveStyle({
-      backgroundColor: "red",
-      color: "blue",
-    });
-    expect(screen.getByLabelText("label")).toBeInTheDocument();
+    expect(screen.getByLabelText("label")).toHaveClass("orbit-kek");
     expect(screen.getByText("badge")).toBeInTheDocument();
     expect(screen.getByTestId("icon")).toBeInTheDocument();
   });
