--- packages/orbit-components/src/tmp-tailwind/primitives/BadgePrimitive/README.md
@@ -1,37 +1,0 @@
-# BadgePrimitive
-
-To implement BadgePrimitive component into your project you'll need to add the import:
-
-```jsx
-import BadgePrimitive from "@kiwicom/orbit-components/lib/primitives/BadgePrimitive";
-```
-
-After adding import into your project you can use it simply like:
-
-```jsx
-<BadgePrimitive>Hello!</BadgePrimitive>
-```
-
-## Props
-
-Table below contains all types of the props available in BadgePrimitive component.
-
-| Name         | Type                    | Default | Description                                                                                                         |
-| :----------- | :---------------------- | :------ | :------------------------------------------------------------------------------------------------------------------ |
-| className    |                         |         | Optional className of the component to allow adding classes that apply styles to text, border and background colors |
-| children     | `React.Node`            |         | The content of the BadgePrimitive.                                                                                  |
-| dataTest     | `string`                |         | Optional prop for testing purposes.                                                                                 |
-| id           | `string`                |         | Set `id` for `BadgePrimitive`                                                                                       |
-| icon         | `React.Node`            |         | The displayed icon on the left.                                                                                     |
-| ariaLabel    | `string`                |         | Adds prop adds `aria-label` to an element, useful for screenreaders.                                                |
-| **carriers** | [`Carrier[]`](#carrier) |         | The content of the CarrierLogo, passed as array of objects.                                                         |
-
-### Carrier
-
-Table below contains all types of the props available for object in Carrier array.
-
-| Name     | Type            | Description                                                                                   |
-| :------- | :-------------- | :-------------------------------------------------------------------------------------------- |
-| **code** | `string`        | The code of the Carrier, defines which logo will be rendered.                                 |
-| name     | `string`        | The name of the Carrier, mainly for information.                                              |
-| type     | [`enum`](#enum) | The preferred placeholder for non-existing carrier. [See Functional specs](#functional-specs) |
