--- packages/orbit-components/src/tmp-tailwind/primitives/BadgePrimitive/index.js.flow
@@ -1,22 +1,0 @@
-// @flow
-/*
-  DOCUMENTATION: https://orbit.kiwi/primitives/badgePrimitive/
-*/
-import * as React from "react";
-import type { StyledComponent } from "styled-components";
-
-import type { Carrier } from "../../../CarrierLogo/index.js.flow";
-import type { Globals } from "../../../common/common.js.flow";
-
-export type Props = {|
-  +children?: React.Node,
-  +icon?: React.Node,
-  +carriers?: Carrier[],
-  +ariaLabel?: string,
-  +className?: string,
-  ...Globals,
-|};
-
-declare export var StyledBadge: StyledComponent<any, any, HTMLElement>;
-
-declare export default React.ComponentType<Props>;
